Stopgap Solution
The bug isn't entirely fixed, it won't be until onground will equal true
when the player is touching a ground object, and false when s/he is not.
But I have changed the movement code so they have a bit more leeway in
midair at best, as opposed to being able to walk without falling.